Changes to logic of building DLL versus static library.
Use just the value of BUILD_SHARED_LIBS to set a define for just Windows DLLs since that is the only context in which a define would be needed and don't define anything for static library.

Removed debug postfix as it is not needed because everything is put in its own directory.

Include files are always the same so do not put them in the build sub-directories.

Removed the DEPENDENCIES variable, it was premature but also, it should not be needed because CMake adds libraries if you say they are public/interface.

Changed LIBRARY_PREFIX to MY_LIBRARY_PREFIX to try and signal better that it is for customization.

Changed LIBRARY_INSTALL back to LIBRARY_TYPE to better parallel BUILD_TYPE

TODO: LIBRARY install only relevant to Unix/Apple? so test this build in Linux
TODO: Unsure about the meaning of INSTALL_INTERFACE and BUILD_INTERFACE
TODO: PUBLIC_HEADER install is not useful if I divide the headers into multiple sub-directories, which I definitely plan to do